Height_Estimation_With_Machine_Learning
# Machine Learning Project: Height Prediction

This repository contains the code and documentation for a machine learning project focused on predicting human heights. The goal of this project is to explore various machine learning algorithms and techniques to accurately estimate the height of individuals based on their personal attributes and characteristics.

## Dataset
The dataset used in this project consists of over 200 samples, with each sample containing information such as gender, age, sports history, health problems, snacking habits, sleep patterns, weekly exercise frequency, average family height, climate conditions, and daily sleep duration. The target variable is the height of individuals.

## Project Highlights
- Implemented multiple predictive models, including Linear Regression, Decision Tree Regressor, Random Forest Regressor, Support Vector Regressor (SVR), and K-Nearest Neighbors Regressor.
- Conducted 10-fold cross-validation to evaluate the performance of each model.
- Explored feature selection techniques, feature transformation using PCA, and normalization to assess their impact on model performance.
- Calculated mean accuracy and standard deviation for each model to compare their performance.
- Performed statistical significance tests (t-test) to determine if there are any significant differences in performance among the models.
- Analyzed the importance of features in predicting height using feature importance scores.

## Results and Insights
- Linear Regression achieved an average accuracy of 0.518 with a standard deviation of 0.164.
- Decision Tree Regressor outperformed other models with an average accuracy of 0.794 and a standard deviation of 0.195.
- Random Forest Regressor also demonstrated strong performance with an average accuracy of 0.839 and a standard deviation of 0.139.
- SVR and K-Nearest Neighbors Regressor achieved lower accuracies of 0.487 and 0.577, respectively.
- Feature importance analysis revealed that gender, average family height, weekly exercise frequency, age, and daily sleep duration are the most influential factors in predicting height.
- PCA-transformed data and Min-Max scaled data were also explored, providing additional insights into the impact of feature transformation and normalization on model performance.

This project showcases the application of various machine learning algorithms and techniques in predicting human heights. The results and insights gained can be valuable in understanding the factors that influence height and can potentially be extended to other areas of research or applications involving height prediction.